# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Somya Fashions
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  category: string;
  imageUrl: string;
  images: string[];
  inStock: boolean;
  featured: boolean;
  rating: number;
  reviewCount: number;
  material: string;
  tags: string[];
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  category: string;
  imageUrl: string;
  images: string[];
  inStock: boolean;
  featured: boolean;
  material: string;
  tags: string[];
}

export interface UpdateProductBody {
  name?: string;
  description?: string;
  price?: number;
  originalPrice?: number | null;
  category?: string;
  imageUrl?: string;
  images?: string[];
  inStock?: boolean;
  featured?: boolean;
  material?: string;
  tags?: string[];
}

export interface CategorySummary {
  name: string;
  count: number;
  imageUrl: string;
}

export interface CartItem {
  id: number;
  productId: number;
  product: Product;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  sessionId: string;
  productId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  sessionId: string;
  quantity: number;
}

export interface PlaceOrderBody {
  sessionId: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  city: string;
  pincode: string;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  total: number;
  status: string;
  items: CartItem[];
  createdAt: string;
}

export interface Customer {
  id: number;
  name: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  pincode: string;
  totalOrders: number;
  totalSpent: number;
  createdAt: string;
}

export interface Payment {
  id: number;
  orderId: number;
  method: string;
  status: string;
  amount: number;
  transactionRef?: string | null;
  upiId?: string | null;
  createdAt: string;
}

export type CreatePaymentBodyMethod =
  (typeof CreatePaymentBodyMethod)[keyof typeof CreatePaymentBodyMethod];

export const CreatePaymentBodyMethod = {
  upi: "upi",
  card: "card",
  cod: "cod",
} as const;

export interface CreatePaymentBody {
  orderId: number;
  method: CreatePaymentBodyMethod;
  upiId?: string | null;
  cardLast4?: string | null;
}

export interface ConfirmPaymentBody {
  transactionRef: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  success: boolean;
}

export type AdminOrderItemsItem = { [key: string]: unknown };

export interface AdminOrder {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  city: string;
  pincode: string;
  total: number;
  status: string;
  paymentMethod?: string | null;
  paymentStatus?: string | null;
  items: AdminOrderItemsItem[];
  createdAt: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  totalCustomers: number;
  totalProducts: number;
  pendingOrders: number;
  recentOrders: AdminOrder[];
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  featured?: boolean;
};

export type GetCartParams = {
  sessionId: string;
};