# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Somya Fashions
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Jewellery product operations
  - name: cart
    description: Shopping cart operations
  - name: orders
    description: Order operations
  - name: customers
    description: Customer operations
  - name: payments
    description: Payment operations
  - name: admin
    description: Admin panel operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/categories:
    get:
      operationId: getCategories
      tags: [products]
      summary: Get all product categories with counts
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySummary"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /orders:
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /payments:
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  /payments/{id}/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPaymentBody"
      responses:
        "200":
          description: Confirmed payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized

  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders (admin)
      responses:
        "200":
          description: All orders with customer and payment details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminOrder"

  /admin/orders/{id}/status:
    put:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrder"

  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        category:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        material:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - category
        - imageUrl
        - images
        - inStock
        - featured
        - rating
        - reviewCount
        - material
        - tags
        - createdAt

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        category:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        material:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - description
        - price
        - category
        - imageUrl
        - images
        - inStock
        - featured
        - material
        - tags

    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        category:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        material:
          type: string
        tags:
          type: array
          items:
            type: string

    CategorySummary:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
        imageUrl:
          type: string
      required: [name, count, imageUrl]

    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
      required: [id, productId, product, quantity]

    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required: [items, total, itemCount]

    AddToCartBody:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
      required: [sessionId, productId, quantity]

    UpdateCartItemBody:
      type: object
      properties:
        sessionId:
          type: string
        quantity:
          type: integer
      required: [sessionId, quantity]

    PlaceOrderBody:
      type: object
      properties:
        sessionId:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        city:
          type: string
        pincode:
          type: string
      required:
        - sessionId
        - customerName
        - customerEmail
        - customerPhone
        - address
        - city
        - pincode

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        total:
          type: number
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        createdAt:
          type: string
      required: [id, customerName, customerEmail, total, status, items, createdAt]

    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        pincode:
          type: string
        totalOrders:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
      required: [id, name, email, phone, address, city, pincode, totalOrders, totalSpent, createdAt]

    Payment:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        method:
          type: string
        status:
          type: string
        amount:
          type: number
        transactionRef:
          type: string
          nullable: true
        upiId:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, orderId, method, status, amount, createdAt]

    CreatePaymentBody:
      type: object
      properties:
        orderId:
          type: integer
        method:
          type: string
          enum: [upi, card, cod]
        upiId:
          type: string
          nullable: true
        cardLast4:
          type: string
          nullable: true
      required: [orderId, method]

    ConfirmPaymentBody:
      type: object
      properties:
        transactionRef:
          type: string
      required: [transactionRef]

    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required: [password]

    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        success:
          type: boolean
      required: [token, success]

    AdminOrder:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        city:
          type: string
        pincode:
          type: string
        total:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
          nullable: true
        paymentStatus:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
        createdAt:
          type: string
      required: [id, customerName, customerEmail, customerPhone, address, city, pincode, total, status, items, createdAt]

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, processing, shipped, delivered, cancelled]
      required: [status]

    AdminStats:
      type: object
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        totalCustomers:
          type: integer
        totalProducts:
          type: integer
        pendingOrders:
          type: integer
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrder"
      required: [totalOrders, totalRevenue, totalCustomers, totalProducts, pendingOrders, recentOrders]